moved search bar to the newcalendar component.
