+++ remindit/src/Components/Utility/NewCalendar.jsx
@@ -65,6 +65,85 @@ const Container = styled('div')({
     height: '100%',
 })
 
+const SearchBar = styled('div')(({ theme }) => ({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center', // Center horizontally
+    marginBottom: '20px',
+    textAlign: 'center', // Center text within input
+}));
+const SearchInput = styled('input')(({ theme }) => ({
+    width: '300px',
+    padding: '8px',
+    borderRadius: '4px',
+    border: `1px solid ${theme.palette.primary.main}`,
+    marginRight: '10px',
+}));
+
+const SearchButton = styled('button')(({ theme }) => ({
+    padding: '8px 16px',
+    borderRadius: '4px',
+    background: theme.palette.primary.main,
+    color: theme.palette.primary.contrastText,
+    border: 'none',
+    cursor: 'pointer',
+}));
+
+const FilterSection = styled('div')(({ theme }) => ({
+    padding: '10px',
+    background: theme.palette.primary.border,
+    borderRadius: '4px',
+    color: theme.palette.primary.contrastText,
+}));
+
+const FilterLabel = styled('label')({
+    display: 'block',
+    marginBottom: '5px',
+    fontWeight: 'bold',
+});
+
+const FilterInput = styled('input')({
+    width: '100%',
+    padding: '8px',
+    borderRadius: '4px',
+    marginBottom: '10px',
+});
+
+const FilterButton = styled('button')(({ theme }) => ({
+    padding: '8px 16px',
+    borderRadius: '4px',
+    background: theme.palette.primary.main,
+    color: theme.palette.primary.contrastText,
+    border: 'none',
+    cursor: 'pointer',
+}));
+
+// Create a new component for the filter section
+function FilterSectionComponent({ keyword, selectedCategory, handleKeywordChange, handleCategoryChange, handleFilterSubmit }) {
+    return (
+        <div>
+            <FilterLabel>Keyword:</FilterLabel>
+            <div style={{ display: 'flex', alignItems: 'center' }}>
+                <FilterInput
+                    type="text"
+                    placeholder="Enter keyword"
+                    value={keyword}
+                    onChange={handleKeywordChange}
+                    style={{ flex: 1, marginRight: '8px' }}
+                />
+                <FilterButton onClick={handleFilterSubmit}>Apply Filter</FilterButton>
+            </div>
+            <FilterLabel>Category:</FilterLabel>
+            <select value={selectedCategory} onChange={handleCategoryChange}>
+                <option value="">All</option>
+                <option value="category1">Category 1</option>
+                <option value="category2">Category 2</option>
+                <option value="category3">Category 3</option>
+            </select>
+        </div>
+    );
+}
+
 function NewCalendar({ date, setDate }) {
     const theme = useTheme();
 
@@ -91,6 +170,43 @@ function NewCalendar({ date, setDate }) {
     const [uploadedAttachment, setUploadedAttachment] = useState(null);
     const [notes, setNotes] = useState('');
     const [notesAttachments, setNotesAttachments] = useState([]);
+    const [showFilter, setShowFilter] = useState(false); // State to manage filter section visibility
+    const [searchTerm, setSearchTerm] = useState('');
+    const [keyword, setKeyword] = useState('');
+    const [selectedCategory, setSelectedCategory] = useState('');
+    const [filteredReminders, setFilteredReminders] = useState([]); // Initialize an empty array for filtered reminders
+
+
+    const handleCategoryChange = (e) => {
+        setSelectedCategory(e.target.value);
+    };
+
+    const handleSearch = () => {
+        // Perform the search functionality here
+        const filteredReminders = reminders.filter(reminder =>
+            reminder.title.toLowerCase().includes(searchTerm.toLowerCase())
+        );
+        setFilteredReminders(filteredReminders);
+
+        // Hide the filter section after search
+        setShowFilter(false);
+    };
+
+    const toggleFilter = () => {
+        setShowFilter(!showFilter); // Toggle filter section visibility
+    };
+
+    const handleKeywordChange = (e) => {
+        setKeyword(e.target.value);
+    };
+
+    const handleFilterSubmit = () => {
+        // Perform filtering based on selected options
+        // You can access the keyword, selectedDate, and selectedCategory here
+        // Example:
+        console.log('Keyword:', keyword);
+        console.log('Selected Category:', selectedCategory);
+    };
 
 
     const closePopup = () => {
@@ -249,6 +365,30 @@ function NewCalendar({ date, setDate }) {
 
     return (
         <Container>
+            <SearchBar onClick={toggleFilter}>
+                <SearchInput
+                    type="text"
+                    placeholder="Search..."
+                    value={searchTerm}
+                    onChange={(e) => setSearchTerm(e.target.value)}
+                    list="suggestions"
+                />
+                <SearchButton onClick={handleSearch}>Search</SearchButton>
+            </SearchBar>
+            {showFilter && (
+                <FilterSectionComponent
+                    keyword={keyword}
+                    selectedCategory={selectedCategory}
+                    handleKeywordChange={handleKeywordChange}
+                    handleCategoryChange={handleCategoryChange}
+                    handleFilterSubmit={handleFilterSubmit}
+                />
+            )}
+            <datalist id="suggestions">
+                {reminders.map((reminder) => (
+                    <option key={reminder.id} value={reminder.title} />
+                ))}
+            </datalist>
             <Top>
                 <Left>
                     <ButtonGroup variant="contained" aria-label="outlined primary button group">
